fix: Resolve all clippy warnings and formatting issues
Clippy fixes:
- Remove needless lifetime parameters from trait implementations
- Change &Vec<T> to &[T] for better API design
- Use std::mem::size_of_val() instead of manual slice size calculations
- Use .div_ceil() instead of manual ceiling division
- Remove unnecessary type casts (u64 as u64)
- Use std::f32::consts::PI and std::f64::consts::E instead of approximations
- Add type annotations where needed for type inference
- Add #[allow(unused_mut)] for cross-platform compatibility

All clippy warnings resolved with -D warnings flag.
All code formatted with cargo fmt.

+++ benches/large_memory.rs
@@ -13,10 +13,13 @@ fn bench_large_memory(c: &mut Criterion) {
         let data: Vec<u64> = (0..num_elements).collect();
 
         let mut group = c.benchmark_group(format!("memory_{}MB", size_mb));
-        group.throughput(Throughput::Bytes((num_elements * 8) as u64));
+        group.throughput(Throughput::Bytes(num_elements * 8));
         group.sample_size(10); // Fewer samples for large data
 
-        println!("\n=== Testing {} MB ({} elements) ===", size_mb, num_elements);
+        println!(
+            "\n=== Testing {} MB ({} elements) ===",
+            size_mb, num_elements
+        );
 
         // Limcode POD
         group.bench_function("limcode_pod_ser", |b| {
