+++ benches/massive_scale.rs
@@ -8,10 +8,10 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion, Throughpu
 fn bench_parallel_massive(c: &mut Criterion) {
     // Test sizes optimized for 240-core machines
     let sizes = [
-        1_000_000,      // 1M elements - threshold for parallelism
-        10_000_000,     // 10M elements - 100 chunks @ 100K each
-        100_000_000,    // 100M elements - 1000 chunks
-        1_000_000_000,  // 1B elements - 10K chunks (svm.run scale!)
+        1_000_000,     // 1M elements - threshold for parallelism
+        10_000_000,    // 10M elements - 100 chunks @ 100K each
+        100_000_000,   // 100M elements - 1000 chunks
+        1_000_000_000, // 1B elements - 10K chunks (svm.run scale!)
     ];
 
     for size in sizes {
@@ -19,8 +19,8 @@ fn bench_parallel_massive(c: &mut Criterion) {
         let size_mb = (size * 8) / 1_000_000;
 
         let mut group = c.benchmark_group(format!("massive_{}M_elements", size / 1_000_000));
-        group.throughput(Throughput::Bytes((size * 8) as u64));
-        group.sample_size(10);  // Fewer samples for huge datasets
+        group.throughput(Throughput::Bytes(size * 8));
+        group.sample_size(10); // Fewer samples for huge datasets
 
         // Single-threaded baseline
         group.bench_function("serial", |b| {
@@ -34,7 +34,11 @@ fn bench_parallel_massive(c: &mut Criterion) {
 
         group.finish();
 
-        println!("\n=== {} million elements ({} MB) ===", size / 1_000_000, size_mb);
+        println!(
+            "\n=== {} million elements ({} MB) ===",
+            size / 1_000_000,
+            size_mb
+        );
         println!("Cores available: {}", rayon::current_num_threads());
     }
 }
